fix(guardrails): adopt freezeClock + canonical tmpdir + test-file caps in shell-audit tests (PR #2374 feedback)
